Refactor AnalyzeColumnCommand for reuse in other places
